fix: mark styled components as client
