+++ frontend/src/components/Loading.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import styled from "@emotion/styled";
 
 const LoaderContainer = styled.div`
