+++ frontend/src/components/Tooltip.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { ReactNode } from "react";
 import styled from "@emotion/styled";
