+++ frontend/src/types/emotion.d.ts
@@ -1,0 +1,13 @@
+import "@emotion/react";
+
+declare module "@emotion/react" {
+    export interface Theme {
+        colors: {
+            accent: string;
+            dark: string;
+            light: string;
+            medium: string;
+        };
+    }
+}
+
